Server/requirements.txt
argon2-cffi==21.3.0
Flask==2.2.2
PyJWT==2.6.0
pymongo==4.3.3
